RustEnvironmentCheck: Check Version if specified
RustEnvironmentCheck currently only checks that the necessary tooling
exists on the system. The only version check it performs is for the
compiler version. This update adds support for repository owners to
optionally require specific versions for cargo-make and cargo-tarpaulin
by setting the required version in the rust-toolchain.toml file at the
workspace root.

Example:
```
[toolchain]
channel = "1.73.0"

[tool]
cargo-tarpaulin = "0.27.3"
cargo-make = "0.37.9"
```